renames .jsx to .js filed and removes the extension from the imports
