+++ frontend/src/components/BetaAccess/SecretBetaAccessGenerator.js
