+++ frontend/src/components/containers/Panel.js
