+++ frontend/src/components/icons/LabelIllustrated.js
