+++ frontend/src/components/user_esteem/LeaderBoard/Leaderboard.js
@@ -1,4 +1,4 @@
-import LeaderBoardRow from './LeaderBoardRow.jsx';
+import LeaderBoardRow from './LeaderBoardRow';
 
 /**
  * LeaderBoard component
