+++ frontend/src/router/router.js
@@ -2,15 +2,14 @@ import store from '../store/store.js';
 import { Router, Route, IndexRoute, browserHistory } from 'react-router';
 import { Provider } from 'react-redux';
 import { syncHistoryWithStore, routerReducer } from 'react-router-redux';
-import UserEsteem from '../components/user_esteem/UserEsteem.jsx';
-import LeaderBoard from '../components/user_esteem/LeaderBoard/LeaderBoard.jsx';
+import UserEsteem from '../components/user_esteem/UserEsteem';
 import AddAppraisee from '../components/user_esteem/AddAppraisee/AddAppraisee';
-import Layout from '../components/global/Layout.jsx';
-import Signup from '../components/signup/Signup.jsx';
-import Signin from '../components/signin/Signin.jsx';
-import ErrorPage from '../components/global/ErrorPage.jsx';
-import Home from '../components/global/Home.jsx';
-import SecretBetaAccessGenerator from '../components/BetaAccess/SecretBetaAccessGenerator.jsx';
+import Layout from '../components/global/Layout';
+import Signup from '../components/signup/Signup';
+import Signin from '../components/signin/Signin';
+import ErrorPage from '../components/global/ErrorPage';
+import Home from '../components/global/Home';
+import SecretBetaAccessGenerator from '../components/BetaAccess/SecretBetaAccessGenerator';
 import * as actionCreators from '../actions/action.js';
 import * as approvalsActionCreators from '../actions/approvalsActions.js';
