+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test:front": "mocha --compilers js:babel-core/register frontend/test --require ignore-styles --watch --recursive",
     "webpack": "webpack --progress --colors --watch",
     "dev": "webpack-dev-server --progress --inline  --history-api-fallback --hot --colors --content-base frontend/",
     "api": "nodemon api/routes/routes.js --harmony_destructuring",
@@ -13,12 +13,14 @@
   "license": "ISC",
   "devDependencies": {
     "babel-core": "^6.5.2",
-    "babel-eslint": "^5.0.0",
+    "babel-eslint": "^6.0.2",
     "babel-loader": "^6.2.3",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
     "bootstrap": "^3.3.6",
     "bootstrap-webpack": "0.0.5",
+    "chai": "^3.5.0",
+    "enzyme": "^2.4.1",
     "eslint": "^2.9.0",
     "eslint-config-airbnb": "^8.0.0",
     "eslint-config-airbnb-base": "^1.0.3",
@@ -29,9 +31,12 @@
     "expose-loader": "^0.7.1",
     "express-livereload": "0.0.24",
     "extract-text-webpack-plugin": "^1.0.1",
+    "ignore-styles": "^4.0.0",
     "imports-loader": "^0.6.5",
     "less": "^2.6.1",
     "less-loader": "^2.2.3",
+    "react-addons-test-utils": "^15.3.1",
+    "react-dom": "^0.14.8",
     "redux-devtools": "^3.1.1",
     "redux-devtools-dock-monitor": "^1.1.0",
     "redux-devtools-log-monitor": "^1.0.5",
@@ -59,6 +64,7 @@
     "react": "^0.14.7",
     "react-bootstrap": "^0.29.2",
     "react-dom": "^0.14.7",
+    "react-dropzone": "^3.5.3",
     "react-redux": "^4.4.0",
     "react-router": "^2.0.0",
     "react-router-redux": "^4.0.0",
